backend/requirements.txt
asgiref==3.5.2
pytz==2022.2.1
sqlparse==0.4.2
Django==3.2.15
Pillow==9.2.0
django-colorfield==0.7.2
djangorestframework==3.13.1
django-filter==22.1
djangorestframework-simplejwt==4.8
djoser==2.1.0
